[BACKLOG-5070] Lineage analyzer fixes for old and new Text File Input steps Fixes two of the integration-test errors by changing the validation KTRs to use the deprecated step ID and moves the analyzers for the old step into an "oldtextfileinput" package.
