+++ core/src/it/resources/repo/validation/file_to_table.ktr
@@ -583,7 +583,7 @@
 
   <step>
     <name>Sacramento crime stats 2006 file</name>
-    <type>TextFileInput</type>
+    <type>OldTextFileInput</type>
     <description/>
     <distribute>Y</distribute>
     <custom_distribution/>
