+++ core/src/test/java/org/pentaho/metaverse/analyzer/kettle/step/oldtextfileinput/TextFileInputStepAnalyzerTest.java
@@ -20,7 +20,7 @@
  *
  ******************************************************************************/
 
-package org.pentaho.metaverse.analyzer.kettle.step.textfileinput;
+package org.pentaho.metaverse.analyzer.kettle.step.oldtextfileinput;
 
 import org.junit.Before;
 import org.junit.Test;
